Split TextGenerator
