--- Sources/SwiftMistralCoreML/SwiftMistralCoreML.swift
@@ -1,2 +1,0 @@
-// The Swift Programming Language
-// https://docs.swift.org/swift-book
